+++ Assets/ChessPieces/Blank.cs
@@ -3,6 +3,6 @@ using Chess.Engine.Bases;
 
 namespace Chess.Assets.ChessPieces;
 
-public class Blank(char symbol, ConsoleColor color, byte x, byte y) : BaseUIObject(symbol, color, x, y)
+public class Blank(char symbol, ConsoleColor color, int x, int y) : BaseUIObject(symbol, color, x, y)
 {
 }
