+++ Scenes/Gameplay/BoardCursor.cs
@@ -4,7 +4,7 @@ using Chess.Engine.Bases;
 
 namespace Chess.Scenes.Gameplay;
 
-public class BoardCursor(char symbol, ConsoleColor color, byte x, byte y) : BaseCursor(symbol, color, x, y)
+public class BoardCursor(char symbol, ConsoleColor color, int x, int y) : BaseCursor(symbol, color, x, y)
 {
     public ChessPiece? SelectedPiece { get; set; } = null;
     public bool IsHoldingChessPiece { get; set; } = false;
